+++ Sources/KeyboardKit/_Deprecated/EmojiKeyboardButton.swift
@@ -1,0 +1,19 @@
+//
+//  EmojiKeyboardButton.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2021-12-19.
+//  Copyright © 2021-2023 Daniel Saidi. All rights reserved.
+//
+
+import SwiftUI
+
+/**
+ This view can be used in an ``EmojiKeyboard`` to display an
+ emoji button based on a certain style.
+
+ Note that since the button only has a tap action, it should
+ not be used in emoji keyboards where a more complex gesture
+ set should be used.
+ */
+