+++ Sources/KeyboardKit/_Deprecated/KeyboardContext+Deprecated.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-public extension KeyboardContext {
-
-    @available(*, deprecated, message: "This is no longer used")
-    static var tempIsPreviewMode: Bool = false
-}