+++ Sources/KeyboardKit/_Deprecated/LocalizedKeyboardLayoutProvider.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-@available(*, deprecated, message: "This will be removed in KeyboardKit 8.0. Use KeyboardLayoutProvider & LocalizedService instead")
-public typealias LocalizedKeyboardLayoutProvider = KeyboardLayoutProvider & LocalizedService