+++ Sources/KeyboardKit/_Deprecated/Styling+Deprecated.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-@available(*, deprecated, renamed: "KeyboardFont.FontType")
-public typealias KeyboardFontType = KeyboardFont.FontType
-
-@available(*, deprecated, renamed: "KeyboardFont.FontWeight")
-public typealias KeyboardFontWeight = KeyboardFont.FontWeight